requirements.txt
absl-py==0.7.1
astor==0.7.1
contextlib2==0.5.5
cycler==0.10.0
dm-sonnet==1.32
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mock==3.0.5
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.1
scipy==1.2.1
semantic-version==2.6.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
tqdm==4.32.1
Werkzeug==0.15.4
wrapt==1.11.1
